+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Helper tasks to:
+  - Update `HelpSet`'s XML `lang` attribute based on the locale in the file name.
+  - Copy common help data for localized help.
 
 ## [0.6.0] - 2021-06-09
 ### Added